Public booking page module #50

+++ packages/api/src/modules/business/usecases/update.business.gallery.photos.usecase.impl.ts
@@ -48,19 +48,25 @@ export class UpdateBusinessGalleryPhotosUseCaseImpl implements UpdateBusinessGal
 
     const oldGalleryPhotos = business.brandApperanceDetails.gallaryPhotosUrls;
 
-    const updated = business.updateGalleryPhotos(command.galleryPhotosKeys);
+    const newGalleryPhotoKeys = command.galleryPhotosKeys.map((value) =>
+      this.s3Service.extractKeyFromUrl(value),
+    );
+
+    const updated = business.updateGalleryPhotos(newGalleryPhotoKeys);
 
     const updateResult = await this.businessRepository.update(updated);
 
     if (!updateResult.ok) return err(updateResult.error as InfrastructureError);
 
     const keysToDelete = oldGalleryPhotos.filter(
-      (key) => !command.galleryPhotosKeys.includes(key),
+      (key) => !newGalleryPhotoKeys.includes(key),
     );
 
-    await Promise.allSettled(
-      keysToDelete.map((key) => this.s3Service.deleteFile(key)),
-    );
+    if (keysToDelete.length > 0) {
+      await Promise.all(
+        keysToDelete.map((key) => this.s3Service.deleteFile(key)),
+      );
+    }
 
     return ok(updated);
   }
